+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Fix undelete reuse and dataservices [#3141](https://github.com/opendatateam/udata/pull/3141)
 
 ## 9.1.4 (2024-08-26)